--- @types/interfaces.d.ts
@@ -1,8 +1,0 @@
-export type CustomFieldParseValueFn = (value: string) => Array<unknown>;
-
-export type CustomFieldFormatValueFn = (inputValues: Array<unknown>) => string;
-
-export interface CustomFieldI18n {
-  parseValue: CustomFieldParseValueFn;
-  formatValue: CustomFieldFormatValueFn;
-}
