+++ src/vaadin-custom-field-mixin.d.ts
@@ -1,3 +1,5 @@
+import { CustomFieldI18n } from './interfaces';
+
 export { CustomFieldMixin };
 
 declare function CustomFieldMixin<T extends new (...args: any[]) => {}>(base: T): T & CustomFieldMixinConstructor;
@@ -21,27 +23,25 @@ interface CustomFieldMixin {
    *
    * The object has the following JSON structure:
    *
-   *          {
-   *            // A function to format given `Array` as
-   *            // component value. Array is list of all internal values
-   *            // in the order of their presence in the DOM
-   *            // This function is called each time the internal input
-   *            // value is changed.
-   *            formatValue: inputValues => {
-   *              // returns a representation of the given array of values
-   *              // in the form of string with delimiter characters
-   *            },
+   * {
+   *   // A function to format given `Array` as
+   *   // component value. Array is list of all internal values
+   *   // in the order of their presence in the DOM
+   *   // This function is called each time the internal input
+   *   // value is changed.
+   *   formatValue: inputValues => {
+   *     // returns a representation of the given array of values
+   *     // in the form of string with delimiter characters
+   *   },
    *
-   *            // A function to parse the given value to an `Array` in the format
-   *            // of the list of all internal values
-   *            // in the order of their presence in the DOM
-   *            // This function is called when value of the
-   *            // custom field is set.
-   *            parseValue: value => {
-   *              // returns the array of values from parsed value string.
-   *            }
+   * // A function to parse the given value to an `Array` in the format
+   * // of the list of all internal values
+   * // in the order of their presence in the DOM
+   * // This function is called when value of the
+   * // custom field is set.
+   * parseValue: value => {
+   *   // returns the array of values from parsed value string.
+   * }
    */
   i18n: CustomFieldI18n;
 }
-
-import { CustomFieldI18n } from '../@types/interfaces';
