+++ vaadin-custom-field.d.ts
@@ -1,2 +1,2 @@
 export * from './src/vaadin-custom-field.js';
-export * from './@types/interfaces';
+export * from './src/interfaces';
